+++ src/app/shared/shared.module.ts
@@ -3,18 +3,25 @@ import { CommonModule } from '@angular/common';
 
 import { HighlightDirective } from './directives/highlight.directive';
 import { BorderDirective } from './directives/border.directive';
+import { OrderByPipe } from './pipes/order-by.pipe';
+import { FormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [
     HighlightDirective,
-    BorderDirective
+    BorderDirective,
+    OrderByPipe
   ],
   imports: [
-    CommonModule
+    CommonModule,
+    FormsModule
   ],
   exports: [
     HighlightDirective,
-    BorderDirective
+    BorderDirective,
+    OrderByPipe,
+    CommonModule,
+    FormsModule
   ]
 })
 export class SharedModule { }